Use natural numbers for test, experiment and trial IDs

+++ benchflow-experiment-manager/application/src/main/java/cloud/benchflow/experimentmanager/tasks/running/CheckTerminationCriteriaTask.java
@@ -28,7 +28,7 @@ public class CheckTerminationCriteriaTask
     logger.info("running - " + experimentID);
 
     int numTrials = experimentModelDAO.getNumTrials(experimentID);
-    int numExectuedTrials = experimentModelDAO.getNumExecutedTrials(experimentID);
+    long numExectuedTrials = experimentModelDAO.getNumExecutedTrials(experimentID);
 
     // TODO - add confidence interval (read termination criteria)
